pyproject.toml
[project]
name = "ttt"
version = "0.1.0"
description = "Diffusion-based posterior sampling for inverse problems (DAPS & DPS)"
requires-python = ">=3.10"
dependencies = [
    # ML framework
    "torch>=2.0.0",
    "torchvision>=0.15.0",

    # Diffusion & generative models
    "diffusers>=0.32.2",
    "transformers>=4.46.3",
    "safetensors>=0.5.3",
    "huggingface-hub>=0.29.3",
    "torchdiffeq>=0.2.5",

    # Configuration
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "pyyaml>=6.0",

    # Image processing & vision
    "pillow>=10.3.0",
    "imageio>=2.34.2",
    "imageio-ffmpeg>=0.5.1",
    "kornia>=0.7.3",
    "einops>=0.8.1",
    "matplotlib>=3.7.5",

    # Evaluation metrics
    "piq>=0.8.0",
    "prettytable>=3.10.0",

    # Scientific computing
    "numpy>=1.24.4",
    "scipy>=1.10.1",

    # Utilities & logging
    "tqdm>=4.64.1",
    "wandb>=0.17.0",
    "setproctitle>=1.3.3",
    "gdown>=5.2.0",

    # CLIP (from git)
    "clip @ git+https://github.com/openai/CLIP.git@main",
]

[project.optional-dependencies]
deblur = [
    "opencv-python>=4.8.0",
    "lmdb>=1.0.0",
    "tensorboard>=1.15.0",
    "ninja>=1.10.0",
]

[tool.uv]
# torch with CUDA is best installed from the PyTorch index.
# Use this to install with CUDA 12.4 support:
#   uv sync --extra-index-url https://download.pytorch.org/whl/cu124
# For CPU-only:
#   uv sync --extra-index-url https://download.pytorch.org/whl/cpu
